# api.yaml
openapi: 3.0.3
info:
  title: Viking Demo API
  version: 1.0.0
  description: Демонстрационное REST API для получения списка созданных викингов

servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки

paths:
  /api/vikings:
    get:
      tags:
        - Vikings
      summary: Получить список всех созданных викингов
      description: Возвращает список всех викингов, созданных в приложении на текущий момент
      operationId: getAllVikings
      responses:
        '200':
          description: Список викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viking'
    post:
      tags:
        - Vikings
      summary: Создать конкретного викинга
      description: Создаёт викинга с параметрами, переданными в теле запроса
      operationId: createViking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viking'
      responses:
        '201':
          description: Викинг успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'

  /api/vikings/{id}:
    put:
      tags:
        - Vikings
      summary: Перезаписать параметры викинга
      description: Полностью обновляет параметры и список снаряжения викинга
      operationId: updateViking
      parameters:
        - $ref: '#/components/parameters/VikingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viking'
      responses:
        '200':
          description: Викинг успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
        '404':
          description: Викинг не найден
    delete:
      tags:
        - Vikings
      summary: Удалить викинга
      description: Удаляет викинга и связанное с ним снаряжение
      operationId: deleteViking
      parameters:
        - $ref: '#/components/parameters/VikingId'
      responses:
        '204':
          description: Викинг успешно удалён
        '404':
          description: Викинг не найден

  /api/vikings/test:
    get:
      tags:
        - Vikings
      summary: Получить список тестовых викингов
      description: Получить список тестовых викингов
      operationId: getTest
      responses:
        '200':
          description: Список викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: string
                  
components:
  parameters:
    VikingId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int32
      description: Идентификатор викинга

  schemas:
    Viking:
      type: object
      required:
        - name
        - age
        - heightCm
        - hairColor
        - beardStyle
        - equipment
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
          example: 1
        name:
          type: string
          example: Ragnar
        age:
          type: integer
          format: int32
          example: 34
        heightCm:
          type: integer
          format: int32
          example: 182
        hairColor:
          $ref: '#/components/schemas/HairColor'
        beardStyle:
          $ref: '#/components/schemas/BeardStyle'
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentItem'

    EquipmentItem:
      type: object
      required:
        - name
        - quality
      properties:
        name:
          type: string
          example: Axe
        quality:
          $ref: '#/components/schemas/EquipmentQuality'

    HairColor:
      type: string
      enum:
        - Blond
        - Red
        - Brown
        - Black
        - Gray
      example: Blond

    BeardStyle:
      type: string
      enum:
        - SHORT
        - LONG
        - BRAIDED
        - FORKED
        - CLEAN_SHAVEN
      example: BRAIDED

    EquipmentQuality:
      type: string
      enum:
        - Common
        - Uncommon
        - Rare
        - Legendary
      example: Rare